refactor: move FieldRequest and derived classes to FieldRequests namespace
Moved FieldRequest and its derived classes (ExistingFieldRequest, NamedFieldRequest, NonExistingFieldRequest) to the PropertyBitPack.SourceGen.Models.FieldRequests namespace. Updated using directives accordingly.
