+++ PropertyBitPack.SourceGen/Models/FieldRequests/FieldRequest.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Text;
 
-namespace PropertyBitPack.SourceGen.Models;
+namespace PropertyBitPack.SourceGen.Models.FieldRequests;
 
 [DebuggerDisplay($"{{{nameof(GetDebuggerDisplay)}(),nq}}")]
 internal class FieldRequest(string name, SpecialType fieldType, bool isExist)
