+++ PropertyBitPack.SourceGen/Models/FieldRequests/NamedFieldRequest.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace PropertyBitPack.SourceGen.Models;
+namespace PropertyBitPack.SourceGen.Models.FieldRequests;
 internal sealed class NamedFieldRequest(string name, SpecialType fieldType) : NonExistingFieldRequest(name, fieldType)
 {
 }
