+++ PropertyBitPack.SourceGen/Models/FieldRequests/NonExistingFieldRequest.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace PropertyBitPack.SourceGen.Models;
+namespace PropertyBitPack.SourceGen.Models.FieldRequests;
 internal class NonExistingFieldRequest(string fieldName, SpecialType specialType) : FieldRequest(fieldName, specialType, false)
 {
 }
